+++ test/autobuild_test.dart
@@ -2,15 +2,17 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:flutter_simple_state/flutter_simple_state.dart';
+import 'package:reactive_state/reactive_state.dart';
 
 void main() {
   testWidgets('observes value change', (WidgetTester tester) async {
     final tagsValue = Value<List<String>>(<String>['a', 'b']);
+    final derivedTags = DerivedValue((get, track) =>
+        List<String>.unmodifiable(<String>['X']..addAll(get(tagsValue))));
     final counter = ValueNotifier<int>(0);
-    final widget = AutoRebuild(builder: (context, get, track) {
+    final widget = AutoBuild(builder: (context, get, track) {
       var count = get(counter);
-      var tags = get(tagsValue);
+      var tags = get(derivedTags);
       return MaterialApp(
         title: 'Test',
         home: Scaffold(
@@ -26,11 +28,11 @@ void main() {
     });
     await tester.pumpWidget(widget);
     expect(find.text('Count: 0'), findsOneWidget);
-    expect(find.text('Tags: a,b'), findsOneWidget);
+    expect(find.text('Tags: X,a,b'), findsOneWidget);
 
     tagsValue.update((t) => t.add('c'));
     await tester.pumpWidget(widget);
-    expect(find.text('Tags: a,b,c'), findsOneWidget);
+    expect(find.text('Tags: X,a,b,c'), findsOneWidget);
 
     counter.value++;
     await tester.pumpWidget(widget);